fix: Fix cirrcular dependency
