+++ packages/mcp-server/src/mcp/sub-server.ts
@@ -1,7 +1,7 @@
 import type { FastifyInstance } from 'fastify';
 import type { McpServer } from '@modelcontextprotocol/sdk/server/mcp.js';
 import FastifyMcpServer, { getMcpDecorator } from 'fastify-mcp-server';
-import type { McpServerName } from '@my-hub/shared/services';
+import type { McpServerName } from '@my-hub/shared/schema';
 import { createHubTokenVerifier } from '../plugins/oauth-verifier.js';
 import { mcpSubServers } from './registry.js';
